NeuralLinearSolve.jl: Neural Network-Based Linear Solver Selection

NeuralLinearSolve.jl uses a pretrained convolutional neural network (CNN) to predict the fastest linear solver for a given sparse matrix. The sparse matrix is converted to a 64x64 spy plot image, which is fed into the CNN to predict the optimal solver from UMFPACK, KLU, and Pardiso.

Installation

To install NeuralLinearSolve.jl, use the Julia package manager:

using Pkg
Pkg.add("NeuralLinearSolve")

Quick Start

using NeuralLinearSolve
using SparseArrays

A = sprand(1000, 1000, 0.01)

# Get the predicted fastest solver
solver = predict_solver(A)

# Get the full probability distribution over solvers
probs = predict_solver_probs(A)

API

NeuralLinearSolve.predict_solver — Function

predict_solver(A::SparseMatrixCSC) -> Symbol

Predict the linear solver expected to be fastest for a sparse matrix.

The bundled convolutional neural network converts A into a fixed-size sparsity image and selects one of the supported solver labels. The model is loaded on the first call and then cached for subsequent predictions.

Arguments

A: Sparse matrix whose sparsity pattern is used for the prediction.

Returns

Symbol: A predicted solver label, currently one of :UMFPACK, :KLU, or :Pardiso.

Examples

julia> using NeuralLinearSolve, SparseArrays

julia> A = sprand(16, 16, 0.2);

julia> predict_solver(A) in (:UMFPACK, :KLU, :Pardiso)
true

NeuralLinearSolve.predict_solver_probs — Function

predict_solver_probs(A::SparseMatrixCSC) -> Dict{Symbol, Float32}

Predict the probability assigned to each supported linear solver for a sparse matrix.

The probabilities are produced by the same cached model used by predict_solver and are useful when downstream code needs to choose among close predictions.

Arguments

A: Sparse matrix whose sparsity pattern is used for the prediction.

Returns

Dict{Symbol, Float32}: A mapping from each supported solver label to its predicted probability.

Examples

julia> using NeuralLinearSolve, SparseArrays

julia> A = sprand(16, 16, 0.2);

julia> probs = predict_solver_probs(A);

julia> isapprox(sum(values(probs)), 1f0)
true

Contributing

Please refer to the SciML ColPrac: Contributor's Guide on Collaborative Practices for Community Packages for guidance on PRs, issues, and other matters relating to contributing to SciML.

See the SciML Style Guide for common coding practices and other style decisions.

There are a few community forums:

The #diffeq-bridged and #sciml-bridged channels in the Julia Slack

The #diffeq-bridged and #sciml-bridged channels in the Julia Zulip

On the Julia Discourse forums

See also SciML Community page

Reproducibility

The documentation of this SciML package was built using these direct dependencies,

Status `~/work/NeuralLinearSolve.jl/NeuralLinearSolve.jl/docs/Project.toml`
  [e30172f5] Documenter v1.17.0
  [653043a4] NeuralLinearSolve v1.0.0-DEV `~/work/NeuralLinearSolve.jl/NeuralLinearSolve.jl`

and using this machine and Julia version.

Julia Version 1.12.7
Commit 6d172b025e4 (2026-08-15 08:05 UTC)
Build Info:
  Official https://julialang.org release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 4 × Intel(R) Xeon(R) 6973P-C
  WORD_SIZE: 64
  LLVM: libLLVM-18.1.7 (ORCJIT, graniterapids)
  GC: Built with stock GC
Threads: 1 default, 1 interactive, 1 GC (on 4 virtual cores)

A more complete overview of all dependencies and their versions is also provided.

Status `~/work/NeuralLinearSolve.jl/NeuralLinearSolve.jl/docs/Manifest.toml`
  [47edcb42] ADTypes v1.24.0
  [a4c015fc] ANSIColoredPrinters v0.0.1
  [621f4979] AbstractFFTs v1.5.0
  [1520ce14] AbstractTrees v0.4.5
  [79e6a3ab] Adapt v4.7.0
  [66dad0bd] AliasTables v1.1.3
  [ec485272] ArnoldiMethod v0.4.0
  [4fba245c] ArrayInterface v7.30.0
  [a9b6321e] Atomix v1.1.3
  [13072b0f] AxisAlgorithms v1.1.0
  [39de3d68] AxisArrays v0.4.8
  [ab4f0b2a] BFloat16s v0.6.1
  [62783981] BitTwiddlingConvenienceFunctions v0.1.6
  [fa961155] CEnum v0.5.0
  [2a0fbf3d] CPUSummary v0.2.7
  [aafaddc9] CatIndices v0.2.2
  [082447d4] ChainRules v1.73.0
  [d360d2e6] ChainRulesCore v1.26.1
  [0b6fb165] ChunkCodecCore v1.0.2
  [4c0bbee4] ChunkCodecLibZlib v1.1.0
  [55437552] ChunkCodecLibZstd v1.0.0
  [fb6a15b2] CloseOpenIntervals v0.1.13
  [aaaa29a8] Clustering v0.15.8
  [da1fd8a2] CodeTracking v3.0.2
  [944b1d66] CodecZlib v0.7.9
  [6b39b394] CodecZstd v0.8.7
  [35d6a980] ColorSchemes v3.31.0
  [3da002f7] ColorTypes v0.12.1
  [c3611d14] ColorVectorSpace v0.11.0
  [5ae59095] Colors v0.13.1
  [bbf7d656] CommonSubexpressions v0.3.1
  [f70d9fcc] CommonWorldInvalidations v1.2.0
  [34da2185] Compat v4.18.1
  [ed09eef8] ComputationalResources v0.3.2
  [187b0558] ConstructionBase v1.6.0
  [150eb455] CoordinateTransformations v0.6.4
  [adafc99b] CpuId v0.3.1
  [dc8bdbbb] CustomUnitRanges v1.0.2
  [9a962f9c] DataAPI v1.16.0
  [864edb3b] DataStructures v0.19.6
  [e2d170a0] DataValueInterfaces v1.0.0
  [8bb1440f] DelimitedFiles v1.9.1
  [163ba53b] DiffResults v1.1.0
  [b552c78f] DiffRules v1.16.0
  [b4f34e82] Distances v0.10.12
  [ffbed154] DocStringExtensions v0.9.5
  [e30172f5] Documenter v1.17.0
  [f151be2c] EnzymeCore v0.8.21
  [4f61f5a4] FFTViews v0.3.2
  [7a1cc6ca] FFTW v1.10.0
  [5789e2e9] FileIO v1.20.0
  [1a297f60] FillArrays v1.17.0
⌅ [53c48c17] FixedPointNumbers v0.8.6
  [587475ba] Flux v0.16.11
  [f6369f11] ForwardDiff v1.4.5
  [d9f16b24] Functors v0.5.3
  [0c68f7d7] GPUArrays v11.5.13
  [46192b85] GPUArraysCore v0.2.0
  [d7ba0133] Git v1.5.0
  [a2bd30eb] Graphics v1.1.3
  [86223c79] Graphs v1.14.0
  [076d061b] HashArrayMappedTries v0.2.0
  [2c695a8d] HistogramThresholding v0.3.1
  [3e5b6fbb] HostCPUFeatures v0.1.18
  [b5f81e59] IOCapture v1.0.0
  [7869d1d1] IRTools v0.4.20
  [615f187c] IfElse v0.1.1
  [2803e5a7] ImageAxes v0.6.12
  [c817782e] ImageBase v0.1.7
  [cbc4b850] ImageBinarization v0.3.1
  [f332f351] ImageContrastAdjustment v0.3.13
  [a09fc81d] ImageCore v0.10.5
  [89d5987c] ImageCorners v0.1.3
  [51556ac3] ImageDistances v0.2.17
  [6a3955dd] ImageFiltering v0.7.12
  [82e4d734] ImageIO v0.6.9
  [6218d12a] ImageMagick v1.4.2
  [bc367c6b] ImageMetadata v0.9.10
  [787d08f9] ImageMorphology v0.4.7
  [2996bd0c] ImageQualityIndexes v0.3.7
  [80713f31] ImageSegmentation v1.10.0
  [4e3cecfd] ImageShow v0.3.8
  [02fcd773] ImageTransformations v0.10.3
  [916415d5] Images v0.26.2
  [9b13fd28] IndirectArrays v1.0.0
  [d25df0c9] Inflate v0.1.5
  [1d092043] IntegralArrays v0.1.6
  [a98d9a8b] Interpolations v0.16.3
  [8197267c] IntervalSets v0.7.14
  [92d709cd] IrrationalConstants v0.2.6
  [c8e1da08] IterTools v1.10.0
  [82899510] IteratorInterfaceExtensions v1.0.0
  [033835bb] JLD2 v0.6.6
  [692b3bcd] JLLWrappers v1.8.0
  [682c06a0] JSON v1.7.1
  [b835a17e] JpegTurbo v0.1.6
  [63c18a36] KernelAbstractions v0.9.42
  [929cbde3] LLVM v9.13.0
  [10f19ff3] LayoutPointers v0.1.17
  [0e77f7df] LazilyInitializedFields v1.3.0
  [8cdb02fc] LazyModules v0.3.1
  [2ab3a3ac] LogExpFunctions v1.0.1
  [bdcacae8] LoopVectorization v0.12.174
  [c2834f40] MLCore v1.1.0
  [7e8f7934] MLDataDevices v1.17.10
  [f1d291b0] MLUtils v0.4.13
  [1914dd2f] MacroTools v0.5.16
  [d125e4d3] ManualMemory v0.1.8
  [dbb5928d] MappedArrays v0.4.3
  [d0879d2d] MarkdownAST v0.1.3
⌅ [626554b9] MetaGraphs v0.8.1
  [e1d29d7a] Missings v1.2.0
  [e94cdb99] MosaicViews v0.3.4
  [872c559c] NNlib v0.9.44
  [77ba4419] NaNMath v1.1.4
  [b8a86587] NearestNeighbors v0.4.29
  [f09324ee] Netpbm v1.1.1
  [653043a4] NeuralLinearSolve v1.0.0-DEV `~/work/NeuralLinearSolve.jl/NeuralLinearSolve.jl`
  [6fe1bfb0] OffsetArrays v1.17.0
  [0b1bfda6] OneHotArrays v0.2.11
  [52e1d378] OpenEXR v0.3.3
  [3bd65402] Optimisers v0.4.9
⌅ [bac558e1] OrderedCollections v1.8.2
  [f57f5aa1] PNGFiles v0.4.5
  [5432bcbf] PaddedViews v0.5.12
⌅ [d96e819e] Parameters v0.12.3
⌅ [69de0a69] Parsers v2.8.7
  [eebad327] PkgVersion v0.3.3
  [1d0040c9] PolyesterWeave v0.2.2
  [f27b6e38] Polynomials v4.1.1
  [aea7be01] PrecompileTools v1.3.4
  [21216c6a] Preferences v1.5.2
  [33c8b6b6] ProgressLogging v0.1.6
  [92933f4c] ProgressMeter v1.11.0
  [43287f4e] PtrArrays v1.4.0
  [4b34888f] QOI v1.0.2
  [94ee1d12] Quaternions v0.7.7
  [b3c3ace0] RangeArrays v0.3.2
  [c84ed2f1] Ratios v0.4.5
  [c1ae055f] RealDot v0.1.0
  [189a3867] Reexport v1.2.2
  [dee08c22] RegionTrees v0.3.4
  [2792f1a3] RegistryInstances v0.1.0
  [ae029012] Requires v1.3.1
  [6038ab10] Rotations v1.7.1
  [fdea26ae] SIMD v3.7.2
  [94e857df] SIMDTypes v0.1.0
  [476501e8] SLEEFPirates v0.6.46
  [431bcebd] SciMLPublic v1.3.0
  [7e506255] ScopedValues v1.6.2
  [efcf1570] Setfield v1.1.2
  [605ecd9f] ShowCases v0.1.0
  [699a6c99] SimpleTraits v0.9.6
  [47aef6b3] SimpleWeightedGraphs v1.5.1
  [45858cf5] Sixel v0.1.5
  [a2af1166] SortingAlgorithms v1.2.3
  [dc90abb0] SparseInverseSubset v0.1.3
  [276daf66] SpecialFunctions v2.9.0
  [cae243ae] StackViews v0.1.2
  [aedffcd0] Static v1.4.6
  [0d7ed370] StaticArrayInterface v1.10.0
  [90137ffa] StaticArrays v1.9.19
  [1e83bf80] StaticArraysCore v1.4.4
  [10745b16] Statistics v1.11.1
  [82ae8749] StatsAPI v1.8.0
  [2913bbd2] StatsBase v0.34.13
  [09ab397b] StructArrays v0.7.3
  [ec057cc2] StructUtils v2.8.5
  [3783bdb8] TableTraits v1.0.1
  [bd369af6] Tables v1.14.0
  [62fd8b95] TensorCore v0.1.1
  [8290d209] ThreadingUtilities v0.5.6
  [731e570b] TiffImages v0.11.9
  [06e1c1a7] TiledIteration v0.5.0
  [3bb67fe8] TranscodingStreams v0.11.3
  [3a884ed6] UnPack v1.0.2
  [013be700] UnsafeAtomics v0.3.2
  [3d5dd08c] VectorizationBase v0.21.74
  [e3aaa7dc] WebP v0.1.3
  [efce3f68] WoodburyMatrices v1.1.0
  [e88e6eb3] Zygote v0.7.12
  [700de1a5] ZygoteRules v0.2.8
  [6e34b625] Bzip2_jll v1.0.9+0
  [2e619515] Expat_jll v2.8.3+0
  [f5851436] FFTW_jll v3.3.12+0
  [61579ee1] Ghostscript_jll v9.55.1+0
⌅ [59f7168a] Giflib_jll v5.2.3+0
  [020c3dae] Git_LFS_jll v3.7.1+0
  [f8c6e375] Git_jll v2.55.0+0
  [c73af94c] ImageMagick_jll v7.1.2029+0
  [905a6f67] Imath_jll v3.2.2+0
  [1d5cc7b8] IntelOpenMP_jll v2025.2.0+0
  [aacddb02] JpegTurbo_jll v3.2.0+1
  [88015f11] LERC_jll v4.1.0+0
  [dad2f222] LLVMExtra_jll v0.0.46+0
  [7e76a0d4] Libglvnd_jll v1.7.1+1
  [94ce4f54] Libiconv_jll v1.18.0+0
  [89763e89] Libtiff_jll v4.7.3+0
  [d3a379c0] LittleCMS_jll v2.19.1+0
  [856f044c] MKL_jll v2025.2.0+0
  [18a262bb] OpenEXR_jll v3.4.14+0
  [643b3616] OpenJpeg_jll v2.5.5+0
  [9bd350c2] OpenSSH_jll v10.5.1+0
  [efe28fd5] OpenSpecFun_jll v0.5.6+0
  [ffd25f8a] XZ_jll v5.8.3+0
  [4f6342f7] Xorg_libX11_jll v1.8.13+0
  [0c0b7dd1] Xorg_libXau_jll v1.0.13+0
  [a3789734] Xorg_libXdmcp_jll v1.1.6+0
  [1082639a] Xorg_libXext_jll v1.3.8+0
  [c7cfdc94] Xorg_libxcb_jll v1.17.1+0
  [c5fb5394] Xorg_xtrans_jll v1.6.0+0
  [3161d3a3] Zstd_jll v1.5.7+1
  [b53b4c65] libpng_jll v1.6.58+0
  [075b6546] libsixel_jll v1.10.5+0
  [c5f90fcd] libwebp_jll v1.6.0+0
  [337d8026] libzip_jll v1.11.3+0
  [1317d2d5] oneTBB_jll v2022.3.0+0
  [0dad84c5] ArgTools v1.1.2
  [56f22d72] Artifacts v1.11.0
  [2a0f44e3] Base64 v1.11.0
  [ade2ca70] Dates v1.11.0
  [8ba89e20] Distributed v1.11.0
  [f43a241f] Downloads v1.7.0
  [7b1f6079] FileWatching v1.11.0
  [9fa8497b] Future v1.11.0
  [b77e0a4c] InteractiveUtils v1.11.0
  [ac6e5ff7] JuliaSyntaxHighlighting v1.12.0
  [4af54fe1] LazyArtifacts v1.11.0
  [b27032c2] LibCURL v0.6.4
  [76f85450] LibGit2 v1.11.0
  [8f399da3] Libdl v1.11.0
  [37e2e46d] LinearAlgebra v1.12.0
  [56ddb016] Logging v1.11.0
  [d6f4376e] Markdown v1.11.0
  [a63ad114] Mmap v1.11.0
  [ca575930] NetworkOptions v1.3.0
  [44cfe95a] Pkg v1.12.1
  [de0858da] Printf v1.11.0
  [3fa0cd96] REPL v1.11.0
  [9a3f8284] Random v1.11.0
  [ea8e919c] SHA v0.7.0
  [9e88b42a] Serialization v1.11.0
  [1a1011a3] SharedArrays v1.11.0
  [6462fe0b] Sockets v1.11.0
  [2f01184e] SparseArrays v1.12.0
  [f489334b] StyledStrings v1.11.0
  [4607b0f0] SuiteSparse
  [fa267f1f] TOML v1.0.3
  [a4e569a6] Tar v1.10.0
  [8dfed614] Test v1.11.0
  [cf7118a7] UUIDs v1.11.0
  [4ec0a83e] Unicode v1.11.0
  [e66e0078] CompilerSupportLibraries_jll v1.3.1+2
  [deac9b47] LibCURL_jll v8.15.0+0
  [e37daf67] LibGit2_jll v1.9.0+0
  [29816b5a] LibSSH2_jll v1.11.3+1
  [14a3606d] MozillaCACerts_jll v2025.11.4
  [4536629a] OpenBLAS_jll v0.3.29+0
  [05823500] OpenLibm_jll v0.8.7+0
  [458c3c95] OpenSSL_jll v3.5.6+0
  [efcefdf7] PCRE2_jll v10.44.0+1
  [bea87d4a] SuiteSparse_jll v7.8.3+2
  [83775a58] Zlib_jll v1.3.1+2
  [8e850b90] libblastrampoline_jll v5.15.0+0
  [8e850ede] nghttp2_jll v1.64.0+1
  [3f19e933] p7zip_jll v17.7.0+0
Info Packages marked with ⌅ have new versions available but compatibility constraints restrict them from upgrading. To see why use `status --outdated -m`

You can also download the manifest file and the project file.